==> scripts/20_pretraitements_dept.Rmd <==
---
title: "Synthèse pêches électriques réseaux `r params$mon_annee`"
subtitle: "Département : `r params$mon_dept`"
author: "OFB - DR Bretagne"
date: "`r format(Sys.time(), 'Le %d %B %Y')`"
output:
  pdf_document: default
  bookdown::word_document2: default
always_allow_html: true
params:
  mon_dept: "22"
  mon_annee: "2021"
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
```




```{r}
# devtools::install_github("pascalirz/aspe")
# devtools::install_github("pascalirz/ggplotExtra")
library(aspe)
library(ggplotExtra)
library(tidyverse)

load(file = "../processed_data/reg_data.RData")
load(file = "../processed_data/mes_depts_geo.RData")
```

```{r}
source(file = "../R/01_fonctions.R")
```


```{r}
dept_ope <- reg_ope %>% 
  filter(dept == params$mon_dept) %>% 
  pull(ope_id) %>% 
  unique()
```

```{r}
dates <- reg_data %>% 
  filter(annee == params$mon_annee) %>% 
  pull(ope_date)

premiere_date <- min(dates) %>% 
  format('%d %B')

derniere_date <- max(dates) %>% 
  format('%d %B %Y')


  
```


Ce Rapport présente l’ensemble des pêches réalisées au titre des réseaux, dans le département, du `r premiere_date` au `r derniere_date`. Les réseaux couverts sont : 

- RCS : Le Réseau de Contrôle de Surveillance est destiné à évaluer l’état général des eaux et à suivre les changements à long terme de l’état des milieux suite à des changements d’origine naturelle ou anthropique.

- RRP : Le Réseau de Référence Pérenne est mis en œuvre pour conforter la connaissance des conditions de référence qui servent à définir le bon état écologique de la DCE.

- RHP : Le Réseau Hydrobiologique Piscicole vise à disposer chaque année d'un état des peuplements piscicoles, à caractériser les variations interannuelles et les tendances, à évaluer les conséquences des évènements naturels exceptionnels et à assurer une veille écologique des espèces.


Dans une premier temps les résultats sont déclinées à une échelle départementale. Les graphiques regroupent l’ensemble des données réunies sur chaque station du `r  params$mon_dept` depuis 1997 (pour les données les plus anciennes). La dimension temporelle permet de mettre en valeur la fluctuation de la qualité de l’eau, de l’ichtyofaune et des paramètres physiques et environnementaux.

Pour finir, les stations sont décrites une à une par la fiche station ASPE. Pour une meilleure lecture, les données sont illustrées par des graphiques permettant de présenter la classe de qualité IPR accompagnée de variables aidant sont interprétation.

```{r}
mon_dept_geo <- mes_depts_geo %>% 
         filter(DEP == params$mon_dept)

mes_pts_dept_geo <- mes_reg_pop_geo %>% 
  sf::st_join(mon_dept_geo) %>% 
  filter(!is.na(DEP))

ggplot(data = mon_dept_geo) +
  geom_sf() +
  geom_sf(data = mes_pts_dept_geo,
          col = "red")
```

# Conditions de pêche

## Caractéristiques environnementales

Ces quatre graphiques présentent l’évolution des paramètres abiotiques environnementaux qui caractérisent la station. La ligne rouge est la médiane de l’ensemble des données des stations du département qui ont été saisies depuis 1997.

Pour l’ensemble de ces graphiques, la courbe représente la médiane des différents paramètres. 

- Granulométrie : L’axe des ordonnées va de 0 à 12. Chacun des chiffres correspond à une référence de granulométrie (voir légende ci-dessous).

0 = Granulométrie inconnue   
1 = Argile (<0,0039mm)   
2 = Limons (0,0039 à 0,0625 mm)   
3 = Sables fins (0,0625 à 2 mm)   
4 = Sables grossier (2 à 5 mm)   
5 = Gravier (2 à 16 mm)   
6 = Cailloux fins (16 à 32 mm)   
7 = Cailloux grossier (32 à 64 mm)   
8 = Pierres fines (64 à 128 mm)   
9 = Pierres grossières (128 à 256 mm)   
10 = Blocs (256 à 1024 mm)   
11 = Rochers (>1 m – substrat immergé avec protubérances)    
12 = Dalles (>1 m – substrat immergé sans protubérances)  

- Largeur mouillée : L’axe des ordonnées est renseigné en mètre (m).
- Longueur station : L'axe des ordonnée est renseigné en mètre (m).
- Profondeur : L’axe des ordonnées est en centimètre (cm).
- Température : L’axe des ordonnées est en degrés Celsius (°C).

  


```{r, fig.width = 9, dpi = 300}
dept_donnees_env <- reg_donnees_env %>% 
  filter(dept == params$mon_dept) 

dept_donnees_env %>% 
  filter(!(variable %in% c("Conductivité", "Intensité", "Tension", "Puissance"))) %>% 
  ggplotExtra::gg_lattice_line(
  var_x = annee,
  var_y = mediane,
  var_lattice = variable,
  threshold_x = 2007,
  lab_y = "Valeur médiane de la variable",
  free_y = TRUE
)
```

## Paramètres électriques

Ces graphiques présentent l’évolution des paramètres liés à l’électricité. La courbe représente la médiane des différents paramètres. La ligne rouge est la médiane des données réalisées sur le département depuis 1997.

- Conductivité : Capacité de l’eau de la station à conduire le courant électrique (charge en ions). Elle se mesure en microsiemens/centimètre (µS/cm).
- Intensité : Intensité délivrée en ampères (A).    
- Puissance : Puissance électrique délivrée en Kilovoltampères (KVA).     
- Tension : Tension électrique délivrée en volts (V).  

```{r, fig.width=9, dpi=300}
dept_donnees_env %>% 
  filter((variable %in% c("Conductivité", "Intensité", "Tension", "Puissance"))) %>% 
  ggplotExtra::gg_lattice_line(
  var_x = annee,
  var_y = mediane,
  var_lattice = variable,
  threshold_x = 2007,
  lab_y = "Valeur médiane de la variable",
  free_y = TRUE
)
```



# IPR

## Tendance départementale

L’indice poissons rivière (IPR) est un indicateur de l’état des peuplements de poissons. Il mesure l’écart entre la composition d’un peuplement de poissons sur une station donnée, observée à partir d’un échantillonnage par pêche électrique, et celle attendue en situation de référence. L’IPR constitue une base standard d’interprétation de résultats d’échantillonnages piscicoles fondés sur l’occurrence et l’abondance des principales espèces d’eau douce présentes en France.



```{r}
dept_ipr <- reg_ope_ipr %>%
  filter(dept == params$mon_dept)

dept_classe_ipr <- classe_ipr %>%
  aspe::ip_completer_classes_couleur()
```

En fond, la couleur correspond aux classes de qualités IPR, depuis « excellent » en bleu jusqu’à « très mauvais » en rouge. 

Les données sont présentées depuis 2007, année de stabilisation des réseaux des stations et du protocole d’échantillonnage. 

Les points correspondent à la médiane des IPR de l’ensemble des stations du `r params$mon_dept`.  La valeur médiane est plus représentative de l’échelle départementale que la moyenne car elle est moins influencée par les valeurs extrêmes qui pourraient intervenir sur quelques stations.

La ligne bleue représente la tendance évolutive de l’IPR médian du département. 

La courbe marron est la médiane régionale de l’IPR. 

La zone grisée délimite les quantiles 25 % et 75 %. Chaque année, un quart des stations a un IPR au-dessus de la zone grisée, la moitié des stations est comprise dans la zone grisée et un quart des stations a un IPR en-dessous.


```{r}
dept_ipr %>%
  group_by(annee) %>%
  summarise(
    ipr25 = quantile(ipr, 0.25),
    ipr50 = quantile(ipr, 0.50),
    ipr75 = quantile(ipr, 0.75)
  ) %>%
  mutate(pop_libelle = params$mon_dept) %>%
  aspe::gg_temp_ipr(var_id_sta = pop_libelle,
                    var_ipr = ipr50,
                    nb_colonnes = 5,
                    df_classes = dept_classe_ipr) +
  geom_ribbon(aes(
    x = annee,
    y = ipr50,
    ymin = ipr25,
    ymax = ipr75
  ),
  alpha = 0.2) +
  geom_smooth(aes(x = annee,
                  y = ipr50),
              method = "lm",
              se = FALSE,
              size = 0.5) +
  ylab("IPR médian") +
  theme(legend.position = "bottom",
        strip.text.x = element_text(size = 10),
        axis.text.x = element_text(angle = 0, hjust = 0.5)) +
  geom_line(data = reg_ope_ipr_mediane,
            aes(x = annee,
                y = ipr_med),
            col = "brown")
```

## Métriques IPR

Le calcul de la note IPR se fait sur 7 métriques, reposant sur les affinités écologiques des poissons et sur leur densité.

En fond, la couleur correspond à la note qualitative de la métrique, en bleu « excellent » rouge « très mauvais ». 

La ligne rouge correspond à la médiane des données récoltées sur l’ensemble des stations depuis 1997 dans le département.

DII = Densité d’individus invertivores (qui se nourrit d’invertébré)   
DIO = Densité d’individu omnivores (mange de tout)   
DIT = Densité d’individus tolérants   
DTI = DEnsité total d'individus   
NTE = Nombre total d’espèces    
NER = Nombre d’espèces rhéophiles (espèces de zones à courant)   
NEL = Nombre d’espèces lithophiles (qui se reproduit sur un substrat rocheux)  




```{r}
dept_metrique <- reg_ope_metrique %>% 
  # filter(ope_id %in% dept_ope) %>% 
  # select(ope_id,
  #        ope_date,
  #        annee,
  #        opi_ner:opi_dti) %>% 
  # select(-ope_date) %>% 
  # pivot_longer(cols = opi_ner:opi_dti,
  #         names_to = "metrique", values_to = "valeur") %>% 
   group_by(annee, metrique) %>% 
    summarise(mediane = median(valeur, na.rm = TRUE)) %>% 
  ungroup() %>% 
  mutate(metrique = str_replace(metrique,
                                pattern = "opi_",
                                replacement = ""),
         metrique = str_to_upper(metrique))
```

```{r, fig.width = 8, fig.height = 7}
ggplotExtra::gg_lattice_ipr(
  df = dept_metrique,
  var_y = mediane,
  var_lattice = metrique,
  metriques = TRUE,
  interactif = FALSE,
  df_classes = dept_classe_ipr
)
```



## IPR par station

En fond, la couleur correspond aux classes de qualités IPR, « excellent » en bleu à « très mauvais » en rouge.

Ce graphique reprend les notes IPR calculées depuis 2007, date à laquelle le protocole a été modifié et appliquer. 

Les points correspondent à la note des IPR de la station par année. 


```{r, fig.width = 10, fig.height = reg_ope_ipr %>% filter(dept == params$mon_dept) %>% pull(pop_libelle) %>% unique %>% length %>% `/`(4) %>% ceiling()%>% `*`(2.5)%>% `+`(1)}

dept_ipr %>%
  gg_temp_ipr(var_id_sta = pop_libelle,
              var_ipr = ipr,
              nb_colonnes = 4,
              df_classes = dept_classe_ipr)
```






# Les espèces

## Les espèces par département

La taille des points indique le pourcentage d’occurrence d'une espèce dans chaque département, c'est à dire le nombre de stations où l'espèce est observée sur le total de stations prospectées dans le département. 

Exemple : Si, dans le Morbihan, le Chevaine a été capturé sur 8 des 16 stations prospectées en `r  params$mon_annee`. Son taux d’occurrence dans ce département est donc de 50%.

Les espèces sont ordonées par nombre de départements de présence, puis par nombre total d’occurrences.

Les écrevisses apparaissent en bleu.


```{r, fig.height = 9, fig.width = 7, dpi = 300}
ggplot(data = reg_dept_1a_pc_oc,
       aes(x = dept_fct,
           y = esp_nom_commun,
           size = pc_occ)) +
  geom_point(col = "darkgreen") +
  scale_size_area(labels = scales::percent) +
  scale_x_discrete(position = 'top') +
  labs(x = "Département",
       y = "",
       size = "Taux d'occurrence") +
  theme(legend.position = "bottom",
        axis.text = element_text(size = 8),
        axis.text.y = element_text(colour = couleurs_ecr),
        legend.text = element_text(size = 8),
        legend.title = element_text(size = 9))
```

ATENTION : Pour une meilleure lisibilité, l’échelle de l’ordonnée est en log. Ainsi, les données ne sont pas écrasées lorsqu’il y a beaucoup ou peu d’individus. 

Les lignes de couleurs correspondent à la densité spécifique moyenne des départements bretons. 

Sur certaines courbes, on observe un pic tous les deux ans. Celui-ci est dû au RCS qui est réalisé un an sur deux. 

L’intérêt de mettre en évidence la densité spécifique moyenne interdépartementale permet de comparer les dynamiques d’évolutions pour chacune des espèces. 



```{r}
# Nombre de taxon par groupe
n_poissons_loc <- length(poissons_loc)

n_eee <- reg_dept_capt %>%
  filter(dept == params$mon_dept,
         esp_nom_commun %in% eee) %>%
  pull(esp_nom_commun) %>%
  unique() %>%
  length()

n_ecr <- reg_dept_capt %>%
  filter(dept == params$mon_dept,
         esp_nom_commun %in% ecr) %>%
  pull(esp_nom_commun) %>%
  unique() %>%
  length()
```




```{r}
# Calcul des dimensions des figures à afficher en fonction du nombre d'espèces dans # chaque groupe.
fig_haut_poissons <- n_poissons_loc %>%
  `/`(6) %>% 
  ceiling() %>% 
  `*`(1.5) %>% 
  `+` (0.5)

fig_haut_ecr <- n_ecr %>%
  `/`(4) %>% 
  ceiling() %>% 
  `*`(1.5) %>% 
  `+` (0.5)

fig_haut_eee <- n_eee %>%
  `/`(4) %>% 
  ceiling() %>% 
  `*`(1.5) %>% 
  `+` (0.5)

fig_larg_eee <- n_eee * 2.35
```



```{r, fig.width = 10, fig.height = fig_haut_poissons}
reg_dept_capt %>% 
  filter(esp_nom_commun %in% poissons_loc) %>% 
  gg_temp_abondance_groupe(var_abondance = densite_moy,
                           var_espece = esp_nom_commun,
                           var_groupe = dept,
                           groupe = params$mon_dept,
                           nb_colonnes = 4,
                           log_axe_y = FALSE)
```

## Abondance numérique

Sur ce premier histogramme, on peut voir le nombre de fois où une espèce a été observée dans le département pour l’année `r  params$mon_annee`. 

Sur le deuxième, c’est la biomasse spécifique totale qui est mise en évidence pour la saison de pêche `r  params$mon_annee`. 

Il est intéressant de mettre en parallèle ces deux tableaux, car cela permet de se rendre compte que ce ne sont pas toujours les espèces les plus abondantes qui représentent le plus gros volume sur l’ensemble des stations et inversement.  


```{r}
reg_dept_capt %>%
  filter(dept == params$mon_dept,
         annee == params$mon_annee,
         effectif > 0) %>% 
  ggplot(aes(x = fct_reorder(esp_code_alternatif2, effectif),
             y = effectif)) +
  geom_bar(stat = "identity",
           fill = "darkgreen") +
  coord_flip() +
  labs(x = "",
       y = "Effectif capturé")
```


## Abondance massique


```{r}
reg_dept_capt %>%
  filter(dept == params$mon_dept,
         annee == params$mon_annee,
         poids > 0) %>% 
  mutate(poids = poids / 1000) %>% 
  ggplot(aes(x = fct_reorder(esp_code_alternatif2, poids),
             y = poids)) +
  geom_bar(stat = "identity",
           fill = "darkgreen") +
  coord_flip() +
  labs(x = "",
       y = "Biomasse capturée (kg)")
```


```{r}
# poids_dept <- reg_dept_poids %>% 
#   filter(dept == params$mon_dept,
#          annee == params$mon_annee)
```

```{r}
# ggplot(data = poids_dept,
#        aes(x = fct_reorder(esp_nom_commun, poids),
#            y = poids)) +
#   geom_bar(stat = "identity",
#            fill = "darkgreen") +
#   coord_flip() +
#   scale_y_continuous(labels = scales::label_number(big.mark = "",
#                                                   accuracy = 1,
#                                                   scale = 1e-3)) +
#   labs(x = "",
#        y = "Biomasse capturée (kg)")
```


## Richesse spécifique

La courbe bleue représente la variation du nombre d’espèces indigènes par an.

En rouge, ce sont les effectifs d’espèces allochtones par an.

Les données précédentes à 2007 sont en pointillé, car les réseaux n'étaient pas stabilisés et le protocole d'échantillonnages différait. 


```{r, fig.height=2.5, fig.width=5}
reg_rs %>% 
  filter(dept == params$mon_dept) %>% 
  ggplot(aes(x = annee,
             y = rs,
             col= espece)) +
  geom_line() + 
  scale_y_continuous(limits = c(0, NA)) +
  labs(y = "Richesse spécifique", x = "", col = "Espèces") + 
  scale_color_brewer(palette = "Set1")
```




## Distribution en taille pour les principales espèces

En inter-annuel, pour les 10 espèces avec les effectifs les plus abondants.

Seules les 10 espèces les plus abondantes sont représentées. Pour calculer des classes de tailles représentatives et discerner les générations, il faut un jeu de donnée suffisamment important.  

Les histogrammes regroupent les effectifs d’une espèce par classe de taille sur une année. En procédant ainsi, on peut identifier différentes générations qui correspondent aux piques (1 an, 1+, 2+). 

En rouge, il s’agit de la médiane, indiquant la taille moyenne qui permet également d’estimer l’âge moyen de la population. 

Remarque : Les classes de tailles sont adaptées pour chaque espèce. Elles ont été définies par … (demander à pascal)


```{r}
# on part des mesures individuelles
longueurs <- reg_mei %>% 
  filter(dept == params$mon_dept,
         mei_taille > 0)

# Histo cumulé pluri-annuel : Les espèces avec mini 1000 individus sur la période
mes_especes <- longueurs %>% 
  group_by(esp_code_alternatif) %>% 
  tally() %>% 
  arrange(-n) %>%
  slice(1:2) %>%
  pull(esp_code_alternatif)

longueurs <- longueurs %>% 
  filter(esp_code_alternatif %in% mes_especes)

```

```{r, fig.height = 8, results = 'hide'}
map(.x = mes_especes,
    .f = gg_dyn_esp,
    df = longueurs,
    var_id_espece = esp_code_alternatif,
    var_id_station = pop_id,
    seuil = 0.01,
    graph_long_mediane = FALSE,
    graph_mois = FALSE,
    graph_effectif = F
)
```



```{r}
save(dept_classe_ipr,
     dept_ipr,
     dept_donnees_env,
     reg_ope_env,
     reg_ope_pp,
     reg_ope_metrique,
     file = "../processed_data/dept_data.RData"
     )
```
